test(workflow): place Linear slug coverage in loader tests
